test:修改react-native-slider的测试用例以及demo
